[Application] Move Provider class to abstract namespace.
