+++ src/Valkyrja/Filesystem/Provider/ComponentProvider.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Filesystem\Provider;
 
 use Override;
 use Valkyrja\Application\Kernel\Contract\ApplicationContract;
-use Valkyrja\Application\Provider\Provider;
+use Valkyrja\Application\Provider\Abstract\Provider;
 
 class ComponentProvider extends Provider
 {
